+++ .idea/misc.xml
@@ -13,6 +13,9 @@
         <entry key="..\:/Users/Alberto/eclipse-workspace/Proyecto UOC/Android-Development-Project/app/src/main/res/layout/activity_postre.xml" value="0.33" />
         <entry key="..\:/Users/Alberto/eclipse-workspace/Proyecto UOC/Android-Development-Project/app/src/main/res/layout/activity_primer_plato.xml" value="0.18426678067550234" />
         <entry key="..\:/Users/Alberto/eclipse-workspace/Proyecto UOC/Android-Development-Project/app/src/main/res/layout/activity_segundo_plato.xml" value="0.18426678067550234" />
+        <entry key="app/src/main/res/layout/activity_first.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/activity_main.xml" value="0.36678200692041524" />
+        <entry key="app/src/main/res/layout/activity_main2.xml" value="0.1" />
       </map>
     </option>
   </component>
