+++ app/src/main/res/layout/activity_first.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".FirstActivity">
+
+    <Button
+        android:id="@+id/bt_Macedonia"
+        android:layout_width="1000px"
+        android:layout_height="200px"
+        android:layout_marginEnd="16dp"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="185dp"
+        android:text="@string/bt_PrimerPlato"
+        android:textSize="24sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+        android:id="@+id/bt_tartaQueso"
+        android:layout_width="1000px"
+        android:layout_height="200px"
+        android:text="@string/bt_SegundoPlato"
+        android:textSize="24sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/bt_Macedonia" />
+
+    <Button
+        android:id="@+id/bt_Churros"
+        android:layout_width="1000px"
+        android:layout_height="200px"
+        android:onClick="postre"
+        android:text="@string/bt_Postres"
+        android:textSize="24sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/bt_tartaQueso" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
