+++ app/src/main/res/layout/activity_main.xml
@@ -7,37 +7,63 @@
     tools:context=".MainActivity">
 
     <Button
-        android:id="@+id/bt_Macedonia"
-        android:layout_width="1000px"
-        android:layout_height="200px"
-        android:layout_marginEnd="16dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="185dp"
-        android:text="@string/bt_PrimerPlato"
-        android:textSize="24sp"
+        android:id="@+id/btnLogin"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:padding="8dp"
+        android:text="Iniciar sesión"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        style="@style/Widget.AppCompat.Button.Colored"
+        />
 
-    <Button
-        android:id="@+id/bt_tartaQueso"
-        android:layout_width="1000px"
-        android:layout_height="200px"
-        android:text="@string/bt_SegundoPlato"
-        android:textSize="24sp"
+    <com.google.android.material.textfield.TextInputLayout
+        android:id="@+id/etPassword"
+        android:layout_width="220dp"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="16dp"
+        android:hint="Password"
+        app:helperText="al menos 8 caracteres"
+        app:endIconMode="password_toggle"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/bt_Macedonia" />
+        app:layout_constraintBottom_toTopOf="@+id/btnLogin">
 
-    <Button
-        android:id="@+id/bt_Churros"
-        android:layout_width="1000px"
-        android:layout_height="200px"
-        android:onClick="postre"
-        android:text="@string/bt_Postres"
-        android:textSize="24sp"
-        app:layout_constraintEnd_toEndOf="parent"
+        <com.google.android.material.textfield.TextInputEditText
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:inputType="textPassword" />
+    </com.google.android.material.textfield.TextInputLayout>
+
+    <com.google.android.material.textfield.TextInputLayout
+        android:id="@+id/etName"
+        android:layout_width="220dp"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="16dp"
+        android:hint="Name"
+        app:endIconMode="clear_text"
+        app:helperText="name and lastname"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/bt_tartaQueso" />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintBottom_toTopOf="@+id/etPassword">
+
+        <com.google.android.material.textfield.TextInputEditText
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:inputType="textPersonName"/>
+    </com.google.android.material.textfield.TextInputLayout>
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Welcome"
+        android:textAppearance="@style/TextAppearance.AppCompat.Large"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toTopOf="@+id/etName"
+        />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
